Add astromatic running code in main function
This relies on the new sex, psfex, sexagain options.
The output files from these commands are directly written
in the right folders.

The connfiguration files for the astromatic suite are packaged
with the entire astro_reduce project, and the code uses
pkg-resources to find the locus of these files on the user's
system at runtime.

Currently, running psfex runs into plotting difficulties,
maybe due the local PLPlot installation

Signed-off-by: Raphaël Duque <duque@iap.fr>